+++ src/App.css
@@ -1,0 +1,5 @@
+html, body {
+  margin: 0;
+  font-family: "Poppins", Arial, Helvetica, sans-serif;
+  background: linear-gradient(#ff2f1305, #ffffff);
+}
