+++ src/Components/WhyBlaze.jsx
@@ -1,0 +1,42 @@
+import "./Stylesheet/whyblaze.css";
+
+const WhyBlaze = () => {
+    return (
+        <section className="whyblaze my-5">
+            <h1>
+                Why Blaze?
+            </h1>
+            <p>Blaze is an open source project, no security issues at all. Get near stock Android feel with all the customizations you need. </p>
+            <div className="parent">
+                <div className="features">
+                    <h3>Open Source</h3>
+                    <p>Being opensource, blaze ensures 100% transpareny with our users. All sources are public and open for everyone to get their hands and eyes on.</p>
+                </div>
+                <div className="features">
+                    <h3>Battery Beast</h3>
+                    <p>Blaze promises improved battery backup, in all of it's supported devices. This happens as our project is clean and free from all bloats providing hustle free experience. </p>
+                </div>
+                <div className="features">
+                    <h3>Fluid Response</h3>
+                    <p>We are commited toward providing our users fast and clean experience, various tweaks are added to give you the blaze you need.</p>
+                </div>
+            </div>
+            <div className="parent">
+                <div className="features">
+                    <h3>User Interface</h3>
+                    <p>Giving you the stock touch from Google, we make sure to provide all the customizations, that one need fall in love with blaze.</p>
+                </div>
+                <div className="features">
+                    <h3>Blaze House</h3>
+                    <p>We are proud to announce about our unique customizations center, here you can toggle between all lovely features from Fonts to Monet.</p>
+                </div>
+                <div className="features">
+                    <h3>Support</h3>
+                    <p>Get instant support for your device on our telegram's community group. Join our telegram community group right now.</p>
+                </div>
+            </div>
+        </section>
+    )
+}
+
+export default WhyBlaze
